🚧 WIP: unstable compiler work (branch support, annotations, goldens)
Made-with: Cursor
