+++ compiler/codegen/testdata/golden/workflow_fan_out.py
@@ -125,28 +125,20 @@ daily = _orca__block("cron",
 class _orca__state_pipeline(TypedDict):
     _orca__trigger: str | None
     _orca__payload: dict | None
-    daily: Any
     researcher: Any
     analyst: Any
     writer: Any
 
-def _orca__node_daily(state: _orca__state_pipeline) -> dict:
-    """Workflow node wrapping 'daily'."""
-    _predecessors = []
-    _input = _orca__gather(state, _predecessors)
-    raise NotImplementedError("workflow node 'daily': block kind 'cron' is not supported in workflows yet")
-    return {"daily": _out}
-
 def _orca__node_researcher(state: _orca__state_pipeline) -> dict:
     """Workflow node wrapping 'researcher'."""
-    _predecessors = ["daily"]
+    _predecessors = []
     _input = _orca__gather(state, _predecessors)
     _out = _orca__invoke_agent(researcher, _input)
     return {"researcher": _out}
 
 def _orca__node_analyst(state: _orca__state_pipeline) -> dict:
     """Workflow node wrapping 'analyst'."""
-    _predecessors = ["daily"]
+    _predecessors = []
     _input = _orca__gather(state, _predecessors)
     _out = _orca__invoke_agent(analyst, _input)
     return {"analyst": _out}
@@ -158,18 +150,18 @@ def _orca__node_writer(state: _orca__state_pipeline) -> dict:
     _out = _orca__invoke_agent(writer, _input)
     return {"writer": _out}
 
-def _orca__route_pipeline(state: _orca__state_pipeline) -> str:
+def _orca__route_pipeline(state: _orca__state_pipeline) -> list[str]:
     """Route to entry node based on trigger source."""
-    return "daily"
+    trigger = state.get("_orca__trigger")
+    if trigger == "daily":
+        return ["researcher", "analyst"]
+    raise ValueError(f"unknown trigger: {trigger!r}")
 
 pipeline = StateGraph(_orca__state_pipeline)
-pipeline.add_node("daily", _orca__node_daily)
 pipeline.add_node("researcher", _orca__node_researcher)
 pipeline.add_node("analyst", _orca__node_analyst)
 pipeline.add_node("writer", _orca__node_writer)
 pipeline.add_conditional_edges(START, _orca__route_pipeline)
-pipeline.add_edge("daily", "researcher")
-pipeline.add_edge("daily", "analyst")
 pipeline.add_edge("researcher", "writer")
 pipeline.add_edge("analyst", "writer")
 pipeline.add_edge("writer", END)
@@ -180,7 +172,6 @@ if __name__ == "__main__":
     initial_state: _orca__state_pipeline = {
         "_orca__trigger": "",
         "_orca__payload": payload,
-        "daily": "",
         "researcher": "",
         "analyst": "",
         "writer": "",
