+++ compiler/helper/helper.go
@@ -1,11 +1,12 @@
-// Package helper provides shared utility functions used across the compiler.
+// Package helper provides standalone string utilities used across the
+// compiler. By design, this package depends only on the Go standard library
+// — no other compiler module — so it can be imported from anywhere without
+// risk of cycles.
 package helper
 
 import (
 	"strings"
 	"unicode"
-
-	"github.com/thakee/orca/compiler/ast"
 )
 
 // ToPascalCase converts a snake_case identifier to PascalCase.
@@ -29,15 +30,4 @@ func ToPascalCase(s string) string {
 		b.WriteString(string(runes))
 	}
 	return b.String()
-}
-
-// HasAnnotation reports whether annotations contains a decorator with the
-// given name (the identifier after @, without the @ prefix).
-func HasAnnotation(annotations []*ast.Annotation, name string) bool {
-	for _, ann := range annotations {
-		if ann != nil && ann.Name == name {
-			return true
-		}
-	}
-	return false
 }
